add server TS

+++ server/controllers/userController.ts
@@ -2,77 +2,103 @@ const userService = require('../service/user-service.ts')
 const bcrypt = require('bcrypt')
 const { validationResult } = require('express-validator')
 const jwt = require('jsonwebtoken')
-const { secret } = require('../config')
-const User = require('../models/user-model')
+const { secret } = require('../config.ts')
+const User = require('../models/user-model.ts')
+import { Response } from 'express';
 
-const generateAccessToken = (id) => {
+const generateAccessToken = (id:string) => {
     const payload = {
         id
     }
     return jwt.sign(payload, secret, { expiresIn: "24h" })
 }
 
+type RegistrationRequestBody = {
+    _id?: string,
+    email: string;
+    password: string;
+    userName: string;
+    userAge: number;
+    interestsAndPreferences: string[];
+   
+};
+
+type ReqFile = Express.Multer.File & { path: string };
+
+export interface TypedRequestBody<T> extends Express.Request {
+    body: T
+    file?: ReqFile
+    query?: any
+    cookies: string
+    
+}
+
 class UserController {
 
-    async registration(req, res, next) {
+    async registration(req: TypedRequestBody<RegistrationRequestBody>, res: Response, next: Function) {
         try {
             const errors = validationResult(req);
             if (!errors.isEmpty()) {
-                return res.status(400).json({ message: 'Ошибка при регистрации', errors })
+                return console.log('Ошибка при регистрации')
             }
-            const { email, password, userName, userAge, interestsAndPreferences } = req.body;
 
-            const userData = await userService.registration(email, password, userName, userAge, interestsAndPreferences);
-            res.cookie('refreshToken', userData.refreshToken, { maxAge: 30 * 24 * 60 * 60 * 1000, httpOnly: true })
-            return res.json(userData)
+            if (req.body) {
+                const { email, password, userName, userAge, interestsAndPreferences } = req.body;
+                const userData = await userService.registration(email, password, userName, userAge, interestsAndPreferences);
 
+                return res.json(userData)          
+            }
         } catch (e) {
             console.log(e)
-            res.send({ message: "Server error" })
         }
     };
 
-    async setAvatar(req, res) {
+    async setAvatar(req: TypedRequestBody<RegistrationRequestBody>, res:Response) {
         try {
             const userId = req.body._id;
             const file = req.file;
-            const { path } = file;
+            if (file) {
+                const { path } = file;
+                const updatedUser = await User.findByIdAndUpdate(
+                    userId,
+                    {
+                        avatar: path,
+                        avatarPath: path
+                    },
+                    { new: true }
+                );
+    
+                if (!updatedUser) {
+                    throw new Error('Пользователь не найден');
+                }
+    
+                await updatedUser.save();
+    
+                res.json(updatedUser);
+            }
+            // const { path } = file;
             if (!file) {
                 throw new Error('Файл не загружен');
             }
-            const updatedUser = await User.findByIdAndUpdate(
-                userId,
-                {
-                    avatar: path,
-                    avatarPath: path
-                },
-                { new: true }
-            );
-
-            if (!updatedUser) {
-                throw new Error('Пользователь не найден');
-            }
-
-            await updatedUser.save();
-
-            res.json(updatedUser);
+            
         } catch (error) {
             console.error('Произошла ошибка при сохранении аватара:', error);
         }
     }
 
-    async deleteUser(req, res) {
+    async deleteUser(req: TypedRequestBody<{ email: string }>, res: Response) {
         try {
-            const { email } = req.body
-            await userService.deleteUser(email)
-            
+            if (req.body) {
+                const { email } = req.body
+                await userService.deleteUser(email)
+            }
+
         } catch (error) {
             console.log(error)
-            res.send({ message: "Error delete user" })
         }
     }
 
-    async recoverPassword(req, res) {
+    async recoverPassword(req: TypedRequestBody<{ email: string }>, res: Response) {
         try {
             const { email } = req.body;
             const newPassord = await userService.recoverPassword(email);
@@ -84,7 +110,7 @@ class UserController {
         }
     };
 
-    async loginTwo(req, res, next) {
+    async loginTwo(req: TypedRequestBody<{ email: string, password: string }>, res: Response, next:Function) {
         try {
             const { email, password } = req.body;
             const userData = await userService.login(email, password);
@@ -95,7 +121,7 @@ class UserController {
         }
     }
 
-    async login(req, res, next) {
+    async login(req: TypedRequestBody<{ username: string, password: string }>, res: Response, next:Function) {
         try {
             const { username, password } = req.body;
             const user = await User.findOne({ username })
@@ -113,25 +139,25 @@ class UserController {
         }
     };
 
-    async logout(req, res, next) {
+    async logout(req: TypedRequestBody<{ refreshToken: string }>, res: Response) {
         try {
-            const { refreshToken } = req.cookies;
+            const refreshToken  = req.cookies;
             const token = await userService.logout(refreshToken)
             res.clearCookie('refreshToken')
             return res.json(token)
         } catch (e) {
-            next(e)
+            console.log(e);
         }
     };
-    
-    async getUserinDatabase(req, res, next) {
+
+    async getUserinDatabase(req: TypedRequestBody<{ email: string, password: string }>, res: Response) {
         try {
-            const email  = req.query.email;
+            const email = req.query.email;
             console.log('email in getUser', email);
             const userData = await userService.getUser(email);
             console.log(userData);
             return res.json(userData)
-            
+
         } catch (error) {
             console.log(error);
             return res.status(500).json({ error: 'Ошибка сервера' });
